moves validation to validators.js
