+++ utils/validators.js
@@ -1,3 +1,24 @@
+const LANGUAGES = {
+  fi: 1,
+  en: 6,
+}
+
+const isValidRow = (row) => {
+  if (!isValidStudentId(row[0])) {
+    return false
+  }
+  if (row[1] && !isValidGrade(row[1])) {
+    return false
+  }
+  if (row[2] && !isValidCreditAmount(row[2])) {
+    return false
+  }
+  if (row[3] && !LANGUAGES[row[3]]) {
+    return false
+  }
+  return true
+}
+
 const isValidStudentId = (id) => {
   if (/^0[12]\d{7}$/.test(id)) {
     // is a 9 digit number with leading 01 or 02
@@ -5,24 +26,22 @@ const isValidStudentId = (id) => {
     const checksum = id
       .substring(1, 8)
       .split('')
-      .reduce((sum, curr, index) => {
-        return (sum + curr * multipliers[index]) % 10
-      }, 0)
+      .reduce((sum, curr, index) => (sum + curr * multipliers[index]) % 10, 0)
     return (10 - checksum) % 10 == id[8]
   }
   return false
 }
 
-const isValidOodiDate = (date) =>
-  /^(3[01]|[12][0-9]|[1-9])\.(1[0-2]|[1-9])\.20[0-9][0-9]$/.test(date) // valid format 29.5.2019
+const isValidOodiDate = date => /^(3[01]|[12][0-9]|[1-9])\.(1[0-2]|[1-9])\.20[0-9][0-9]$/.test(date) // valid format 29.5.2019
 
-const isValidGrade = (grade) => /^([0-5]|Hyv\.|Hyl\.)$/.test(grade) // 0 to 5, Hyv. or Hyl.
+const isValidGrade = grade => /^([0-5]|Hyv\.|Hyl\.)$/.test(grade) // 0 to 5, Hyv. or Hyl.
 
-const isValidCreditAmount = (credits) => /^[0-9]?[0-9],[05]$/.test(credits) // 0,0 to 99,5 in 0,5 steps
+const isValidCreditAmount = credits => /^[0-9]?[0-9],[05]$/.test(credits) // 0,0 to 99,5 in 0,5 steps
 
 module.exports = {
   isValidStudentId,
   isValidOodiDate,
   isValidGrade,
-  isValidCreditAmount
+  isValidCreditAmount,
+  isValidRow,
 }
